LayoutTests/platform/glib/imported/w3c/web-platform-tests/selection/caret/move-around-generated-content-expected.txt

abc

one

three
xyz

paragraph before the blockquote

quote
after the blockquote


PASS Initializing tests
PASS Moving caret between list items using arrow keys
PASS Moving caret past the block-quote using arrow keys
PASS Moving caret between list items using arrow keys: caret should start at the beginning of the first line
FAIL Moving caret between list items using arrow keys: arrow-down should move the caret to the start of the first list item assert_equals: expected "(#text \"one\", 0)" but got "(#text \"abc\", 0)"
FAIL Moving caret between list items using arrow keys: arrow-down should move the caret to the second list item assert_equals: expected "(<li>, 0)" but got "(#text \"abc\", 0)"
FAIL Moving caret between list items using arrow keys: arrow-down should move the caret to the third list item assert_equals: expected "(#text \"three\", 0)" but got "(#text \"abc\", 0)"
FAIL Moving caret between list items using arrow keys: arrow-down should move the caret to the final paragraph assert_equals: expected "(#text \"xyz\", 0)" but got "(#text \"abc\", 0)"
PASS Moving caret past the block-quote using arrow keys: caret should start at the beginning of the first line
FAIL Moving caret past the block-quote using arrow keys: arrow-down should move the caret to the start of the block quote assert_equals: expected "(#text \"quote\", 0)" but got "(#text \"paragraph before the blockquote\", 0)"
FAIL Moving caret past the block-quote using arrow keys: arrow-down should move the caret to the final paragraph assert_equals: expected "(#text \"after the blockquote\", 0)" but got "(#text \"paragraph before the blockquote\", 0)"
FAIL Moving caret past the block-quote using arrow keys: arrow-left should move the caret to the end of the quote assert_equals: expected "(#text \"quote\", 5)" but got "(#text \"paragraph before the blockquote\", 0)"
FAIL Moving caret past the block-quote using arrow keys: arrow-up should move the caret to into the first line assert_equals: expected "(#text \"paragraph before the blockquote\", 10)" but got "(#text \"paragraph before the blockquote\", 0)"
FAIL Moving caret past the block-quote using arrow keys: arrow-right should move the caret forward by one character assert_equals: expected "(#text \"paragraph before the blockquote\", 11)" but got "(#text \"paragraph before the blockquote\", 0)"
FAIL Moving caret past the block-quote using arrow keys: arrow-down should move the caret to the end of the quote assert_equals: expected "(#text \"quote\", 5)" but got "(#text \"paragraph before the blockquote\", 0)"
